fix(ventura): move Ventura component to components
