+++ package.json
@@ -49,6 +49,7 @@
     "firebase-tools": "^8.1.1",
     "identity-obj-proxy": "^3.0.0",
     "jest": "^25.4.0",
+    "ncp": "^2.0.0",
     "prettier": "^2.0.4",
     "react": "^16",
     "react-dom": "^16",
