+++ src/index.ts
@@ -3,5 +3,5 @@ import './theme.css';
 export { default as Button } from './components/Button/Button';
 export { default as Input } from './components/Input/Input';
 export { default as Radio } from './components/Radio/Radio';
-export { default as Ventura } from './gatsby-theme-docz/components/Ventura/Ventura';
+export { default as Ventura } from './components/Ventura/Ventura';
 export * from 'react-feather';
